Send UNIX timestamp along with JS touch events, instead of systemUptime
Summary:
We want to be able to instrument touch processing delays in JS, which does not have access to systemUptime; therefore we want a UNIX timestamp, which JS has access to and can compare to the touch time.

It only matters that there is relative consistency between multiple touch events in JS, which is still the case; so this should have no impact on product code.

Changelog: [Internal]

Reviewed By: yungsters

Differential Revision: D26705429

fbshipit-source-id: 0f2db726048fcab9a30e830970d7d8a8d2eae446